Fixed sender index assertion. Fixed stats initialization and handle case where no message has been received

+++ internal/azsvcbus/azsvcbus.go
@@ -190,9 +190,9 @@ func ( azSvcBus *AzSvcBus )receiveMessage( idx int ) {
             }
 
             if senderIdxPropVal, exists := message.ApplicationProperties[ idxPropName ]; exists {
-                senderIdx, ok := senderIdxPropVal.( int )
+                senderIdx, ok := senderIdxPropVal.( int64 )
                 if ok {
-                    azSvcBus.stats.UpdateReceiverStat( idx, senderIdx, 1, uint64( helpers.GetCurTimeStamp( ) - msgInst.TimeStamp ) )
+                    azSvcBus.stats.UpdateReceiverStat( idx, int( senderIdx ), 1, uint64( helpers.GetCurTimeStamp( ) - msgInst.TimeStamp ) )
                 } else {
                     glog.Errorf( "%v: Invalid sender index in message application properties", id )
                 }
